requirements.txt
slack-sdk>=3.27
apscheduler>=3.10
python-dotenv>=1.0
requests>=2.31
openai>=1.0
